# src/python/dart/web/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /action/{action_id}/subscription/elements:
    get:
      operationId: listActionSubscriptionElements
      summary: "get the subscription elements for an action"
      tags:
        - "Action"
        - "SubscriptionElement"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
      responses:
        "200":
          description: "the subscription elements"
          schema:
            $ref: '#/definitions/PagedSubscriptionElementsResponse'
        "400":
          description: "if the action doesn't have a subscription"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /dataset/{dataset_id}/subscription:
    post:
      operationId: createDatasetSubscription
      summary: "create a subscription to a dataset"
      tags:
        - "Dataset"
        - "Subscription"
      parameters:
        - name: "dataset_id"
          type: "string"
          in: "path"
          required: true
        - name: "subscription"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/SubscriptionResponse"
...
  /subscription:
    get:
      operationId: listSubscriptions
      summary: "list subscriptions"
      tags:
        - "Subscription"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "the subscriptions"
          schema:
            $ref: '#/definitions/PagedSubscriptionsResponse'
  /subscription/{subscription_id}:
    get:
      operationId: getSubscription
      summary: "get a subscription"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "the subscription"
          schema:
            $ref: '#/definitions/SubscriptionResponse'
    put:
      operationId: updateSubscription
      summary: "update a subscription"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
        - name: subscription
          in: body
          required: true
          schema:
            $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "the subscription"
          schema:
            $ref: '#/definitions/SubscriptionResponse'
        "400":
          description: "subscription is in an invalid state to update"
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: patchSubscription
      summary: "update a subscription using a JSON Patch"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: "the subscription"
          schema:
            $ref: '#/definitions/SubscriptionResponse'
        "400":
          description: "subscription is in an invalid state to update"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteSubscription
      summary: "delete a subscription"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/OKResponse"
  /subscription/{subscription_id}/element_stats:
    get:
      operationId: getSubscriptionElementStats
      summary: "get the element statistics for a subscription"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "the element statistics"
          schema:
            $ref: '#/definitions/ObjectsResponse'
  /subscription/{subscription_id}/elements:
    get:
      operationId: listSubscriptionElements
      summary: "list the elements for a subscription"
      tags:
        - "Subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
        - name: state
          in: query
          type: string
        - name: processed_after_s3_path
          in: query
          type: string
        - name: limit
          in: query
          type: integer
        - name: offset
          in: query
          type: integer
      responses:
        "200":
          description: "the subscription elements"
          schema:
            $ref: '#/definitions/PagedSubscriptionElementsResponse'
  /trigger:
    post:
      operationId: createTrigger
      summary: "create a new trigger"
      tags:
        - "Trigger"
      parameters:
        - name: trigger
          in: body
          required: true
          schema:
            $ref: "#/definitions/Trigger"
      responses:
        "200":
          description: "the trigger"
          schema:
            $ref: '#/definitions/TriggerResponse'
    get:
      operationId: listTriggers
      summary: "list triggers"
      tags:
        - "Trigger"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "the triggers"
          schema:
            $ref: '#/definitions/PagedTriggersResponse'
  /trigger/{trigger_id}:
    get:
      operationId: getTrigger
      summary: "get a trigger"
      tags:
        - "Trigger"
      parameters:
        - name: trigger_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "the trigger"
          schema:
            $ref: '#/definitions/TriggerResponse'
    put:
      operationId: updateTrigger
      summary: "update a trigger"
      tags:
        - "Trigger"
      parameters:
        - name: trigger_id
          in: path
          type: string
          required: true
        - name: trigger
          in: body
          required: true
          schema:
            $ref: "#/definitions/Trigger"
      responses:
        "200":
          description: "the trigger"
          schema:
            $ref: '#/definitions/TriggerResponse'
    patch:
      operationId: patchTrigger
      summary: "patch a trigger"
      tags:
        - "Trigger"
      parameters:
        - name: trigger_id
          in: path
          type: string
          required: true
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: "the trigger"
          schema:
            $ref: '#/definitions/TriggerResponse'
    delete:
      operationId: deleteTrigger
      summary: "delete a trigger"
      tags:
        - "Trigger"
      parameters:
        - name: trigger_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/OKResponse"
  /trigger_type:
    get:
      operationId: listTriggerTypes
      summary: "get trigger types"
      tags:
        - "TriggerType"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
      responses:
        "200":
          description: "the trigger types"
          schema:
            $ref: '#/definitions/PagedTriggerTypesResponse'
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
...
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
...
  ErrorResponse:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResponse:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...
  ObjectsResponse:
    properties:
      results:
        items:
          type: object
        type: array
    type: object
...
  PagedSubscriptionElementsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/SubscriptionElement'
        type: array
      total:
        type: integer
    type: object
  PagedSubscriptionsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Subscription'
        type: array
      total:
        type: integer
    type: object
  PagedTriggerTypesResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/TriggerType'
        type: array
      total:
        type: integer
    type: object
  PagedTriggersResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Trigger'
        type: array
      total:
        type: integer
    type: object
...
  Subscription:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/SubscriptionData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  SubscriptionData:
    additionalProperties: false
    properties:
      dataset_id:
        type: string
      failed_time:
        type: string
        x-nullable: true
      generating_time:
        type: string
        x-nullable: true
      initial_active_time:
        type: string
        x-nullable: true
      message_id:
        type: string
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      on_failure_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      on_success_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      queued_time:
        type: string
        x-nullable: true
      s3_path_end_prefix_exclusive:
        description: The exclusive s3 path end prefix
        pattern: ^s3://.+$
        type: string
        x-nullable: true
      s3_path_regex_filter:
        description: A regex pattern the s3 path must match
        type: string
        x-nullable: true
      s3_path_start_prefix_inclusive:
        description: The inclusive s3 path start prefix
        pattern: ^s3://.+$
        type: string
        x-nullable: true
      state:
        default: INACTIVE
        enum:
        - INACTIVE
        - QUEUED
        - GENERATING
        - FAILED
        - ACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
    required:
    - name
    - dataset_id
    - user_id
    type: object
  SubscriptionElement:
    type: object
  SubscriptionResponse:
    properties:
      results:
        $ref: '#/definitions/Subscription'
    type: object
  Trigger:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/TriggerData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  TriggerData:
    additionalProperties: false
    properties:
      args:
        type: object
      extra_data:
        type: object
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      state:
        default: INACTIVE
        enum:
        - ACTIVE
        - INACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      trigger_type_name:
        type: string
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
      workflow_ids:
        default: []
        items:
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''workflow_id '' + $index }}'
          type: tabarray
    required:
    - name
    - trigger_type_name
    - args
    - user_id
    type: object
  TriggerResponse:
    properties:
      results:
        $ref: '#/definitions/Trigger'
    type: object
  TriggerType:
    properties:
      description:
        type: string
      name:
        type: string
      params_json_schema:
        type: object
    type: object
...